Update Python code.
Commit only Python source updates and leave non-code assets/config changes out of this push.

+++ Pretrain/train_kernel_script.py
@@ -20,7 +20,7 @@ if __name__ == '__main__':  # pragma: no cover
     #train_kernel(dataset_name = 'pointmaze', specific_dataset ='medium', batch_size = 256, lr = 3e-4, num_steps = 50000, ensemble_size=3, λ_reg=1e-3)
     #train_kernel(dataset_name = 'pointmaze', specific_dataset ='large', batch_size = 256, lr = 3e-4, num_steps = 300000, ensemble_size=3, λ_reg=1e-3)
     dataset = 'cube'
-    specific_dataset = 'double'
+    specific_dataset = 'single'
     train_mog_kernel(
          dataset_name = dataset,
          specific_dataset = specific_dataset,
@@ -47,4 +47,4 @@ if __name__ == '__main__':  # pragma: no cover
                 quantile = 0.99,
                 noise_floor = 5e-4)
 
-                
+
